Adds assignment tables between Workflows/Insights provisioned users and persons (Recidiviz/recidiviz-data#36826)
## Description of the change

In order to enable metric creation for the provisioned user units of
analysis (to unblock full hydration of assignments for all experiment
variants), this PR builds views that determine which users should be
associated with which person_ids. The logic is as follows:
- For workflows users, use the type of access configuration in a given
state (statewide, or district-wide, or officer caseload specific) to
determine the set of persons whom a workflows user could theoretically
view in the Workflows tool. This uses some of the logic originally
developed in this now [deprecated
PR](https://github.com/Recidiviz/recidiviz-data/pull/29270) by creating
a yaml that specifies permissioning by state.
- For insights users, join user emails to state staff and then to
supervision_unit_supervisor_sessions, which maps supervisors to persons
based on the intermediate line staff who is supervised by the supervisor
and who supervises the client.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Advances Recidiviz/recidiviz-data#36180

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

---------

Co-authored-by: mayukas <mayuka@stanford.edu>
GitOrigin-RevId: 484558d2bf9d60527d7e7ad4cf6758798c116c98