Update assessment score state-specific logic (Recidiviz/recidiviz-data#35795)
## Description of the change

Right now, there are both AND and OR conditions at the top level in the
WHERE clause, which means the AND conditions are basically getting
ignored, e.g. allowing null assessment_date values. I moved the
state-specific assessment_type OR conditions into a parentheses section
together so the top-level contains only AND statements.

I used Sam's notebook to check for changes in eligibility and it seems
there are none! However I'd make like a second pair of eyes to check
that I ran this correctly:
https://colab.research.google.com/drive/1Jkbe_fZlF_4OqigqgoFygKrss_9rGcsc#scrollTo=Pj2tZl1-x74i

In assessment_score_sessions itself, it looks like this change would
remove 43,840 records where assessment_date is null and 2,174 records
where both assessment_score and assessment_level are null. No other
records are removed or added.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 61e22e5158d627b0b1619eb20fb0d0e694e810e8