[US_IX][Workflows] Remove treatment court clients as eligible for lsu (Recidiviz/recidiviz-data#37076)
## Description of the change

Based on what clients are being marked as ineligible for and what
officers are requesting, we are removing clients on `DIVERSION` (ie on
specialty treatment courts) from the LSU opportunities.

- Looks like 54/1828 (3%) clients surfaced (eligible or almost eligible)
are on diversion so would be excluded
- Then for the jii opp, 18/342 fewer eligible (5.5%) and then 137/4182
(3%) fewer almost eligible.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 33d183a0b81d4529121700912c8b9b6d9097c2ca