[Aggregated Metrics] Allow for arbitrary period column values (Recidiviz/recidiviz-data#37036)
## Description of the change

In preparation for migrating impact reports agg metrics to the new
aggregated metrics format, updates the `period_name_type` value on
`MetricTimePeriodConfig` to a string `period_name` value that can be any
value.

This will be important for impact reports because we have multiple
configs that, for example, use week-long metrics, but each start at
different start points for the rolling interval.

This PR also renames `MetricTimePeriodConfig.week_periods()` to
`MetricTimePeriodConfig.week_periods_monday_aligned()` to make it
clearer where the weeks start.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Part of Recidiviz/recidiviz-data#35911

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: addaa0234eb8a1c4d3f51b9da27e21421891990e